% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_helpers.R
\docType{data}
\name{.term_transformations}
\alias{.term_transformations}
\title{Internal map of term → transformation type}
\format{
An object of class \code{list} of length 5.
}
\usage{
.term_transformations
}
\description{
Internal map of term → transformation type
}
\keyword{internal}
