% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_helpers.R
\docType{data}
\name{.transformations}
\alias{.transformations}
\title{Internal transformations}
\format{
An object of class \code{list} of length 4.
}
\usage{
.transformations
}
\description{
These define how estimates are transformed before applying
the Wald CI (e.g., taking log or logit).
These define how to back-transform point estimates
Functions take arguments \verb{(eta, sd, z)} and return values
on the original scale.
}
\keyword{internal}
