% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_model_0}
\alias{fit_model_0}
\alias{fit_model_1}
\title{Fit survival models to estimate exposure-specific hazards for G-computation
approach}
\usage{
fit_model_0(
  data,
  outcome_time,
  outcome_status,
  exposure,
  exposure_time,
  covariates,
  formula_0 = NULL
)

fit_model_1(
  data,
  outcome_time,
  outcome_status,
  exposure,
  exposure_time,
  covariates,
  tau,
  censor_time = NULL,
  formula_1 = NULL
)
}
\arguments{
\item{data}{A data frame with one row per individual containing the columns
named in \code{outcome_time}, \code{outcome_status}, \code{exposure}, \code{exposure_time}, and
\code{covariates}. Missing values for all columns except \code{exposure_time} are not allowed.}

\item{outcome_time}{Name of the follow-up time for the outcome of interest, i.e.
time to either the event or right-censoring, whichever occurs first. Time should
be measured from a chosen time origin (e.g. study start, enrollment, or age).}

\item{outcome_status}{Name of the event indicator for the outcome. The underlying column should be
numeric (\code{1} = event, \code{0} = censored).}

\item{exposure}{Name of the exposure indicator. The underlying column should
be numeric (\code{1} = exposed during follow-up, \code{0} = never exposed during
follow-up).}

\item{exposure_time}{Name of the time to exposure, measured on the same time scale
as that used for \code{outcome_time}. Must be a non-missing numeric value exposed individuals
and must be set to \code{NA} for unexposed individuals.}

\item{covariates}{Character vector of covariates to adjust for when fitting
the hazard models. Include all known, measured confounders of
exposure and censoring.  Covariates must be measured or defined at the chosen time origin.}

\item{formula_0}{Optional right hand side of the formula for model 0. By default, uses \code{covariates}.}

\item{tau}{Delay period so that events before tau are not included.}

\item{censor_time}{Time after exposure at which exposed
individuals are censored during model fitting to prevent extrapolation. By default,
no censoring is applied.}

\item{formula_1}{Optional right hand side of the formula for model 1. By default, uses \code{covariates}
plus natural spline of vaccination day (4 df).}
}
\value{
A fitted \code{coxph} object with additional component \verb{$data} containing
the analysis dataset used for fitting:
\itemize{
\item For \code{fit_model_0()}: includes the survival tuple \verb{(Y}, \code{event})  and
covariates adjusted for in model, where \code{Y} is the time from time origin
to first of endpoint, censoring or exposure time (for exposed individuals).
\item For \code{fit_model_1()}: includes the survival tuple \verb{(T1}, \code{event}), \verb{<exposure_time>},
and covariates adjusted for in model, where \code{T1} is the time from exposure to
endpoint or censoring, with additional censoring by \code{censor_time}. Only includes
exposed individuals at risk \code{tau} days after exposure.
}
}
\description{
\code{fit_model_0()} fits a Cox model to estimate risk for unexposed
individuals on the original time scale. Includes all individuals, censoring
exposed individuals at their time of exposure. By default, model is adjusted for by
\verb{<covariates>}, included as simple linear terms.

\code{fit_model_1()} fits a Cox model to estimate risk for exposed individuals on
the time scale of time since exposure. Includes exposed individuals who
remain at risk \code{tau} days after exposure. Individuals are additionally
censored at \code{censor_time} days after exposure to avoid extrapolation beyond
the time period of interest. By default, model is adjusted for
\verb{<covariates>}, included as simple linear terms, and exposure time is included as a
natural cubic spline with 4 degrees of freedom.
}
