% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_marginalizing_weights.R
\name{get_gp}
\alias{get_gp}
\title{Get empirical probability distributions g(d|x) and p(x) based on input data}
\usage{
get_gp(df, outcome_time, exposure, exposure_time, covariates, tau)
}
\arguments{
\item{df}{A data frame representing the target population for the estimated distributions
\itemize{
\item typically, population involves the "analysis-eligible" subset of a given data source
}}

\item{outcome_time}{Name of the follow-up time for the outcome of interest, i.e.
time to either the event or right-censoring, whichever occurs first. Time should
be measured from a chosen time origin (e.g. study start, enrollment, or age).}

\item{exposure}{Name of the exposure indicator. The underlying column should
be numeric (\code{1} = exposed during follow-up, \code{0} = never exposed during
follow-up).}

\item{exposure_time}{Name of the time to exposure, measured on the same time scale
as that used for \code{outcome_time}. Must be a non-missing numeric value exposed individuals
and must be set to \code{NA} for unexposed individuals.}

\item{covariates}{Character vector of covariates to adjust for when fitting
the hazard models. Include all known, measured confounders of
exposure and censoring.  Covariates must be measured or defined at the chosen time origin.}

\item{tau}{Numeric. Only vaccinated who are at-risk tau days after vaccination are included.}
}
\value{
A list of the following:
\describe{
\item{g_weights}{A data frame containing the covariate-conditional day probabilities.
Columns include \code{group_id} identifying the covariate group, \verb{<exposure_time> specifying the day, }prob\verb{specifying the covariate-conditional day probability, and each variables in}covariates\verb{\} \\item\{p_weights\}\{A data frame containing the covariate probabilities. Columns include }group_id\verb{ identifying the covariate group,}prob\verb{specifying the covariate probability, and each variable in}covariates`}
}
}
\description{
Get empirical probability distributions g(d|x) and p(x) based on input data
}
\keyword{internal}
