% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_marginalizing_weights.R
\name{get_observed_gp}
\alias{get_observed_gp}
\title{Estimate the marginalizing distribution in the observed or matched analysis-eligible
populations.}
\usage{
get_observed_gp(data, outcome_time, exposure, exposure_time, covariates, tau)
}
\arguments{
\item{outcome_time}{Name of the follow-up time for the outcome of interest, i.e.
time to either the event or right-censoring, whichever occurs first. Time should
be measured from a given time origin (e.g. study start, enrollment, or age)
for all individuals.}

\item{exposure}{Name of the exposure indicator. The underlying column should
be numeric (\code{1} = exposed during follow-up, \code{0} = never exposed during
follow-up).}

\item{exposure_time}{Name of the time to exposure, measured on the same time scale
as that used for \verb{outcome time}. Set \code{exposure_time} of all unexposed
individuals to \code{NA}.}

\item{covariates}{Character vector of covariates to adjust for when fitting
the hazard models; should include all known confounders of
exposure and censoring.  Covariates must be measured or defined at the chosen time origin.}

\item{tau}{Numeric. Only vaccinated who are at-risk tau days after vaccination are included.}
}
\value{
A list of two data frames containing the marginalizing distributions returned from calls to \code{\link[=get_gp]{get_gp()}}
}
\description{
\code{get_observed_gp} returns marginalizing distributions  based on the original observed data
\code{get_matching_gp} returns marginalizing distributions based on the matched-analysis data
}
\keyword{internal}
