% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_one_nomatch.R
\name{get_one_nomatch}
\alias{get_one_nomatch}
\title{Internal function to compute cumulative incidence and effect measures}
\usage{
get_one_nomatch(
  data,
  outcome_time,
  outcome_status,
  exposure,
  exposure_time,
  covariates,
  tau,
  eval_times,
  censor_time = max(eval_times),
  custom_gp_list = NULL,
  keep_models = TRUE,
  return_gp_list = TRUE
)
}
\arguments{
\item{data}{A data frame with one row per individual containing the columns
named in \code{outcome_time}, \code{outcome_status}, \code{exposure}, \code{exposure_time}, and
\code{covariates}. Missing values for all columns except \code{exposure_time} are not allowed.}

\item{outcome_time}{Name of the follow-up time for the outcome of interest, i.e.
time to either the event or right-censoring, whichever occurs first. Time should
be measured from a chosen time origin (e.g. study start, enrollment, or age).}

\item{outcome_status}{Name of the event indicator for the outcome. The underlying column should be
numeric (\code{1} = event, \code{0} = censored).}

\item{exposure}{Name of the exposure indicator. The underlying column should
be numeric (\code{1} = exposed during follow-up, \code{0} = never exposed during
follow-up).}

\item{exposure_time}{Name of the time to exposure, measured on the same time scale
as that used for \code{outcome_time}. Must be a non-missing numeric value exposed individuals
and must be set to \code{NA} for unexposed individuals.}

\item{covariates}{Character vector of covariates to adjust for when fitting
the hazard models. Include all known, measured confounders of
exposure and censoring.  Covariates must be measured or defined at the chosen time origin.}

\item{eval_times}{Numeric vector specifying the timepoints at which to compute
cumulative incidence and the derived effect measures. The timepoints should
be expressed in terms of time since exposure and use the same units
as that used for \code{outcome_time} and \code{exposure_time} (e.g. days). All values must be greater
than \code{ immune_lag} and and should correspond to clinically meaningful follow-up
durations, such as 30, 60, or 90 days after exposure. A fine grid of
timepoints (e.g., \code{eval_times = (immune_lag + 1):100}) can be provided if cumulative
incidence curves over time are desired.}

\item{censor_time}{Time after exposure at which exposed
individuals are administratively censored during model fitting. Default:
\code{max(eval_times)}. This limits estimation to the observed  period of interest and
prevents extrapolation beyond the largest evaluation time.
Typically, users should leave this at the default.}

\item{keep_models}{Logical; return the two fitted hazard models used to compute
cumulative incidences?
Default: \code{TRUE}.}

\item{return_gp_list}{Logical; return marginalizing weights? Default is
TRUE.}

\item{weights_source}{Character string specifying how to construct marginalizing weights.
One of:
\itemize{
\item \code{"observed"} (default): estimate weights from the observed data;
\item \code{"custom"}: use user-specified weights provided via \code{custom_weights} argument;
}}

\item{return_matching}{Logical; return matched datasets? Default is
TRUE if \code{weights_source = "matched"}. When \code{weights_source != "matched"}, this is
automatically set to \code{FALSE}.}
}
\value{
List with components:
\describe{
\item{pt_estimates}{Matrix of point estimates with one row per evaluation time and
one column per measure (\code{cuminc_0}, \code{cuminc_1}, \code{risk_ratio},\code{vaccine_effectiveness}).}
\item{model_0, model_1}{Fitted Cox models (if \code{keep_models = TRUE})}
\item{gp_list}{Marginalizing distributions (if \code{return_gp_list = TRUE})}
}
}
\description{
This is an internal function that performs the actual estimation
of cumulative incidences and derived effect measures
using the G-computation style approach. It is called by \code{\link[=nomatch]{nomatch()}} and \code{\link[=one_boot_nomatch]{one_boot_nomatch()}}.
Users should typically call these functions rather than calling this function directly.
For historical reasons, this function handles more complex inputs than
exposed in the \code{nomatch()} interface. In particular, it includes
an options to
\itemize{
\item set censor time in hazard model for the exposed
\item weight by the weights from a matched dataset and
\item provide hazard model formulas or prefit objects
}
}
\keyword{internal}
