% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsve-package.R
\docType{package}
\name{nomatch-package}
\alias{nomatch-package}
\title{nomatch: Estimate Effectiveness of Interventions in Target Trial Emulation Observational Studies Without Matching}
\description{
Implements estimators of marginal cumulative incidence for binary interventions using observational data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ewu16.github.io/nomatch/}
  \item \url{https://github.com/ewu16/nomatch}
  \item Report bugs at \url{https://github.com/ewu16/nomatch/issues}
}

}
\author{
\strong{Maintainer}: Emily Wu \email{emily.wu2@emory.edu} (\href{https://orcid.org/0009-0003-5487-6753}{ORCID})

}
\keyword{internal}
