% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_boot_matching.R
\name{one_boot_matching}
\alias{one_boot_matching}
\title{Compute one bootstrap replicate of matching analysis method}
\usage{
one_boot_matching(
  matched_data,
  outcome_time,
  outcome_status,
  exposure,
  exposure_time,
  tau,
  timepoints,
  limit_type = "fixed",
  data = NULL,
  id_name = NULL,
  matching_vars = NULL,
  replace = NULL,
  keep_boot_samples = TRUE
)
}
\arguments{
\item{matched_data}{A data frame for the matched cohort created using
\code{\link[=match_rolling_cohort]{match_rolling_cohort()}}.}

\item{outcome_time}{Name of the follow-up time for the outcome of interest, i.e.
time to either the event or right-censoring, whichever occurs first. Time should
be measured from a chosen time origin (e.g. study start, enrollment, or age).}

\item{outcome_status}{Name of the event indicator for the outcome. The underlying column should be
numeric (\code{1} = event, \code{0} = censored).}

\item{exposure}{Name of the exposure indicator. The underlying column should
be numeric (\code{1} = exposed during follow-up, \code{0} = never exposed during
follow-up).}

\item{exposure_time}{Name of the time to exposure, measured on the same time scale
as that used for \code{outcome_time}. Must be a non-missing numeric value for exposed individuals
and must be set to \code{NA} for unexposed individuals.}

\item{timepoints}{Numeric vector specifying the timepoints at which to compute
cumulative incidence and the derived effect measures. The timepoints should
be expressed in terms of time since exposure and use the same units
as that used for \code{outcome_time} and \code{exposure_time} (e.g. days). All values must be greater
than \code{ immune_lag} and and should correspond to clinically meaningful follow-up
durations, such as 30, 60, or 90 days after exposure. A fine grid of
timepoints (e.g., \code{timepoints = (immune_lag + 1):100}) can be provided if cumulative
incidence curves over time are desired. By default, the sequence from \code{immune_lag + 1} to
the maximum event time in the exposed group, by units of 1, is used.}
}
\value{
The \code{pt_estimates} component returned by \code{\link[=get_one_matching]{get_one_matching()}},
for a bootstrap sample.
}
\description{
Compute one bootstrap replicate of matching analysis method
}
\keyword{internal}
