% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_boot_nomatch.R
\name{one_boot_nomatch}
\alias{one_boot_nomatch}
\title{Compute one bootstrap replicate for G-computation method}
\usage{
one_boot_nomatch(
  data,
  outcome_time,
  outcome_status,
  exposure,
  exposure_time,
  covariates,
  eval_times,
  tau,
  custom_gp_list
)
}
\arguments{
\item{data}{A data frame with one row per individual containing the columns
named in \code{outcome_time}, \code{outcome_status}, \code{exposure}, \code{exposure_time}, and
\code{covariates}. Missing values for all columns except \code{exposure_time} are not allowed.}

\item{outcome_time}{Name of the follow-up time for the outcome of interest, i.e.
time to either the event or right-censoring, whichever occurs first. Time should
be measured from a chosen time origin (e.g. study start, enrollment, or age).}

\item{outcome_status}{Name of the event indicator for the outcome. The underlying column should be
numeric (\code{1} = event, \code{0} = censored).}

\item{exposure}{Name of the exposure indicator. The underlying column should
be numeric (\code{1} = exposed during follow-up, \code{0} = never exposed during
follow-up).}

\item{exposure_time}{Name of the time to exposure, measured on the same time scale
as that used for \code{outcome_time}. Must be a non-missing numeric value exposed individuals
and must be set to \code{NA} for unexposed individuals.}

\item{covariates}{Character vector of covariates to adjust for when fitting
the hazard models. Include all known, measured confounders of
exposure and censoring.  Covariates must be measured or defined at the chosen time origin.}

\item{eval_times}{Numeric vector specifying the timepoints at which to compute
cumulative incidence and the derived effect measures. The timepoints should
be expressed in terms of time since exposure and use the same units
as that used for \code{outcome_time} and \code{exposure_time} (e.g. days). All values must be greater
than \code{ immune_lag} and and should correspond to clinically meaningful follow-up
durations, such as 30, 60, or 90 days after exposure. A fine grid of
timepoints (e.g., \code{eval_times = (immune_lag + 1):100}) can be provided if cumulative
incidence curves over time are desired.}
}
\value{
The \code{pt_estimates} component returned by \code{\link[=get_one_nomatch]{get_one_nomatch()}},
for a bootstrap sample.
}
\description{
This function computes the point estimate for
a bootstrap replicate.
}
\keyword{internal}
