% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{plot_effect_panel}
\alias{plot_effect_panel}
\title{Internal plotting function for effect panel plots}
\usage{
plot_effect_panel(
  plot_data,
  ci_type,
  alpha,
  effect = c("risk_ratio", "risk_difference", "relative_risk_reduction"),
  trt_0_label = "Unexposed",
  trt_1_label = "Exposed",
  colors = NULL
)
}
\arguments{
\item{plot_data}{A data frame containing estimates to plot. Must include
columns: \code{t0} (time), \code{estimate}, \code{term} (one of "cuminc_0", "cuminc_1", "relative_risk_reduction"),
\verb{<ci_type>_lower}, \verb{<ci_type>_upper}, and \code{method}}

\item{ci_type}{Character string specifying the type of confidence interval
band to plot, one of \verb{"wald", "percentile", "simul"}, or \code{"none"}.  Must
choose a \code{ci_type} whose lower and upper bounds are already computed in
\code{estimates} component of \code{x}. Uses \code{"wald"} by default if available.}

\item{effect}{The effect measure to plot next to the cumulative incidence
plots. Either \code{"risk_ratio"}(default), \code{"relative_risk_reduction"}  or
\code{"risk_difference"}.}
}
\description{
Internal plotting function for effect panel plots
}
\keyword{internal}
