% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_psi_dx.R
\name{predict_from_model_0}
\alias{predict_from_model_0}
\alias{predict_from_model_1}
\title{Compute conditional cumulative incidences from fitted Cox models}
\usage{
predict_from_model_0(fit_0, exposure_time, t0, tau, newdata)

predict_from_model_1(fit_1, exposure_time, t0, tau, newdata)
}
\value{
The \code{newdata} data frame with three additional columns:
\itemize{
\item \code{predict_from_model_0()}: \code{surv_0_d_plus_tau}, \code{surv_0_d_plus_t0}, and \code{psi_0_dx}
\item \code{predict_from_model_1()}: \code{surv_1_tau}, \code{surv_1_t0}, and \code{psi_1_dx}
}
}
\description{
These functions compute the conditional day- and covariate-specific cumulative
incidence for each row of \code{newdata}.
\itemize{
\item \code{predict_from_model_0()} computes \eqn{\psi_0(t_0; d, x) = 1 - S_0(d+t_0 \mid x)/S_0(d+\tau \mid x)}
by calling \code{predict(fit_0, newdata, type = "survival")} at eval_times
\code{d + t0} and \code{d + tau}.
}

\code{predict_from_model_1()} computes \eqn{\psi_1(t_0; d, x) = 1 - S_1(t_0 \mid d, x)/S_1(\tau \mid d, x)}
\verb{by calling }predict(fit_1, newdata, type = "survival")\verb{at eval_times}t0\code{and}tau`.
}
\keyword{internal}
