% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vefit_methods.R
\name{print.nomatchfit}
\alias{print.nomatchfit}
\title{Print method for effectiveness fits}
\usage{
\method{print}{nomatchfit}(
  x,
  digits = 3,
  effect = c("risk_ratio", "risk_difference", "relative_risk_reduction"),
  ...
)
}
\arguments{
\item{x}{An object of class \code{nomatchfit} created by \code{\link[=nomatch]{nomatch()}} or \code{\link[=matching]{matching()}}.}

\item{digits}{Integer indicating the number of decimal places to display. Default is 3.}

\item{effect}{The effect measure to output.  Either
\code{"risk_ratio"}(default), \code{"relative_risk_reduction"}  or \code{"risk_difference"}.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
Invisibly returns the input object \code{x}. Called primarily for its side effect
of printing a summary to the console.
}
\description{
Prints a concise summary of effectiveness estimates from a fitted model.
}
