% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_nomatch_population.R
\name{summarize_nomatch_population}
\alias{summarize_nomatch_population}
\title{Estimate covariate means for population represented in \code{nomatch()} estimand}
\usage{
summarize_nomatch_population(data, fit)
}
\arguments{
\item{data}{A data frame. The same data passed to \code{\link[=nomatch]{nomatch()}}.}

\item{fit}{A \code{nomatchfit} object created by \code{\link[=nomatch]{nomatch()}}. Must have been
fit with \code{keep_models = TRUE} (the default) and \code{weights_source = "observed"}.}
}
\value{
A data frame of the summary statistics with columns:
\describe{
\item{variable}{Covariate name}
\item{label}{Factor level label (empty string for numeric variables)}
\item{mean}{Mean (or weighted proportion for factor levels)}
}
}
\description{
Estimate covariate means for the weighted principal strata population
represented in \code{nomatch()} estimand. Since the principal strata populations
are not directly observed, simple summary statistics cannot be used. Instead,
the means are estimated as weighted means where the weights depend on
the both the original estimand weights and on
predicted probabilities of belonging to different principal strata.
Currently only implemented for the case when the estimand weights
represent the observed distribution of exposure days and covariates among the exposed.
}
\examples{
fit <- nomatch(
  data = simdata,
  outcome_time = "Y",
  outcome_status = "event",
  exposure = "V",
  exposure_time = "D_obs",
  covariates = c("x1", "x2"),
  timepoints = seq(30, 180, by = 30),
  immune_lag = 14
)
summarize_nomatch_population(simdata, fit)

}
