% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_one_matching.R
\name{get_one_matching}
\alias{get_one_matching}
\title{Compute VE point estimate in matched data set}
\usage{
get_one_matching(
  matched_data,
  outcome_time,
  outcome_status,
  exposure,
  exposure_time,
  tau,
  timepoints,
  keep_models = TRUE,
  keep_adata = TRUE
)
}
\arguments{
\item{matched_data}{A data frame for the matched cohort created using \code{\link[=match_rolling_cohort]{match_rolling_cohort()}}.}

\item{outcome_time}{Name of the follow-up time for the outcome of interest, i.e.
time to either the event or right-censoring, whichever occurs first. Time should
be measured from a chosen time origin (e.g. study start, enrollment, or age).}

\item{outcome_status}{Name of the event indicator for the outcome. The underlying column should be
numeric (\code{1} = event, \code{0} = censored).}

\item{exposure}{Name of the exposure indicator. The underlying column should
be numeric (\code{1} = exposed during follow-up, \code{0} = never exposed during
follow-up).}

\item{exposure_time}{Name of the time to exposure, measured on the same time scale
as that used for \code{outcome_time}. Must be a non-missing numeric value exposed individuals
and must be set to \code{NA} for unexposed individuals.}

\item{timepoints}{Numeric vector specifying the timepoints at which to compute
cumulative incidence and the derived effect measures. The timepoints should
be expressed in terms of time since exposure and use the same units
as that used for \code{outcome_time} and \code{exposure_time} (e.g. days). All values must be greater
than \code{ immune_lag} and and should correspond to clinically meaningful follow-up
durations, such as 30, 60, or 90 days after exposure. A fine grid of
timepoints (e.g., \code{timepoints = (immune_lag + 1):100}) can be provided if cumulative
incidence curves over time are desired.}

\item{keep_models}{Logical; return the two fitted hazard models used to compute
cumulative incidences?
Default: \code{TRUE}.}
}
\value{
The object returned by \code{\link[=compute_km_ve]{compute_km_ve()}}
}
\description{
First, creates the analysis matched data set based on \code{matched_data}.
Then computes Kaplan Meier estimates of cumulative incidence  on the analysis data set.
}
\keyword{internal}
